tests - don't drop non-existent db in setup
contribution from PR #33 - Svein Helge <svein.helge.grinden@gmail.com>

extended the contribution to cover both test classes

+++ EfEnumToLookupTests/Tests/TestConfiguration.cs
@@ -16,7 +16,10 @@ namespace EfEnumToLookupTests.Tests
 			// Using setup rather than teardown to make it easier to inspect the database after running a test.
 			using (var context = new MagicContext())
 			{
-				context.Database.Delete();
+				if (context.Database.Exists())
+				{
+					context.Database.Delete();
+				}
 			}
 		}
