+++ EfEnumToLookupTests/Tests/TestStuff.cs
@@ -19,7 +19,10 @@ namespace EfEnumToLookupTests.Tests
 			// Using setup rather than teardown to make it easier to inspect the database after running a test.
 			using (var context = new MagicContext())
 			{
-				context.Database.Delete();
+				if (context.Database.Exists())
+				{
+					context.Database.Delete();
+				}
 			}
 
 			Database.SetInitializer(new TestInitializer(new EnumToLookup()));
